requirements.txt
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
fredapi==0.5.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
